+++ chess/fen.cpp
@@ -333,9 +333,9 @@ bool parse_fen_noclock( std::string const& fen, position& p, color::type& c, std
 		p.bitboards[i].b[bb_type::all_pieces] = p.bitboards[i].b[bb_type::pawns] | p.bitboards[i].b[bb_type::knights] | p.bitboards[i].b[bb_type::bishops] | p.bitboards[i].b[bb_type::rooks] | p.bitboards[i].b[bb_type::queens] | p.bitboards[i].b[bb_type::king];
 
 		p.bitboards[i].b[bb_type::pawn_control] = 0;
-		unsigned long long pawns = p.bitboards[i].b[bb_type::pawns];
+		uint64_t pawns = p.bitboards[i].b[bb_type::pawns];
 		while( pawns ) {
-			unsigned long long pawn = bitscan_unset( pawns );
+			uint64_t pawn = bitscan_unset( pawns );
 			p.bitboards[i].b[bb_type::pawn_control] |= pawn_control[i][pawn];
 		}
 	}
